Updated Map - See descripion
I updated Map.java using Justin's GpsTracker class from the web and
created a test visualization in Map.java. The coordinates update every 5
seconds. They seem to be extremely off and we may need to look into the
Google Maps API at least for location. As of right now, the application
is working however, so downloading and testing is doable for the GPS.